Enforcing float32 for all fit inputs to prevent numerical precision errors. Moved linelist tables to float32. Fixed q3dpro [OI]/Ha error. PG1411 cube now fits through.

+++ q3dfit/q3dout.py
@@ -81,9 +81,9 @@ class q3dout:
                  gd_indx=None, fitran_indx=None):
 
         self.fitrange = fitrange
-        self.wave = wave
-        self.spec = spec
-        self.spec_err = spec_err
+        self.wave = np.float32(wave)
+        self.spec = np.float32(spec)
+        self.spec_err = np.float32(spec_err)
 
         self.fitran_indx = fitran_indx
         self.gd_indx = gd_indx
